Add files via upload
It reads how to add teachers. 
If file is selected the user can select the file by navigating through the computer. 
If manual is selected the user need to specify the number of teachers, then add the name of the teacher and subject respectively in very field. If all values are null then error message is shown. if any field is blank then that field is nullified
The values are then written to a CSV file called Teacher